+++ CHANGES.md
@@ -13,6 +13,7 @@
 - Remap references when renaming tokens [Taiga #10202](https://tree.taiga.io/project/penpot/us/10202)
 - Tokens panel nested path view [Taiga #9966](https://tree.taiga.io/project/penpot/us/9966)
 - Improve usability of lock and hide buttons in the layer panel. [Taiga #12916](https://tree.taiga.io/project/penpot/issue/12916)
+- Add renderer selector and JPEG quality control in exports [Github #2425](https://github.com/penpot/penpot/issues/2425)
 
 ### :bug: Bugs fixed